build: Updating mediawiki/mediawiki-phan-config to 0.9.0
Depends-On: I9def9907a6666be7fc916afff9a8552806fce681

+++ includes/Drafts.php
@@ -16,7 +16,7 @@ abstract class Drafts {
 			// Drafts stay forever
 			return 0;
 		}
-		return wfTimestamp( TS_UNIX ) - ( $egDraftsLifeSpan * 60 * 60 * 24 );
+		return (int)wfTimestamp( TS_UNIX ) - ( $egDraftsLifeSpan * 60 * 60 * 24 );
 	}
 
 	/**
